# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /settings/unlock:
    post:
      operationId: unlockSettings
      tags: [settings]
      summary: Exchange the admin password for a short-lived settings token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnlockSettingsInput"
      responses:
        "200":
          description: Unlocked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnlockSettingsResult"
        "401":
          description: Incorrect password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Too many attempts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /settings/azure:
    get:
      operationId: getAzureSettings
      tags: [settings]
      summary: Get Azure integration settings status (secrets masked)
      responses:
        "200":
          description: Azure settings status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AzureSettingsStatus"
    put:
      operationId: updateAzureSettings
      tags: [settings]
      summary: Save Azure integration settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AzureSettingsInput"
      responses:
        "200":
          description: Updated settings status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AzureSettingsStatus"
  /files/{id}/import-document:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    post:
      operationId: importDocument
      tags: [loans]
      summary: Upload a PDF and extract loan/lease data from it
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/DocumentUpload"
      responses:
        "200":
          description: Extraction result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentImportResult"
        "400":
          description: Missing settings or invalid upload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: Document could not be read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /trash/{type}/{id}/restore:
    parameters:
      - name: type
        in: path
        required: true
        schema:
          type: string
          enum: [client, file, loan]
      - name: id
        in: path
        required: true
        schema:
          type: string
    post:
      operationId: restoreTrashItem
      tags: [trash]
      summary: Restore a trashed item (and any trashed parents needed to reach it)
      responses:
        "204":
          description: Restored
        "404":
          description: Item not found in trash
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /trash/{type}/{id}:
    parameters:
      - name: type
        in: path
        required: true
        schema:
          type: string
          enum: [client, file, loan]
      - name: id
        in: path
        required: true
        schema:
          type: string
    delete:
      operationId: purgeTrashItem
      tags: [trash]
      summary: Permanently delete a trashed item (cannot be undone)
      responses:
        "204":
          description: Permanently deleted
        "404":
          description: Item not found in trash
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  schemas:
...
    AzureSettingsStatus:
      type: object
      properties:
        docIntelEndpoint:
          type: string
        openaiEndpoint:
          type: string
        openaiDeployment:
          type: string
        docIntelKeySet:
          type: boolean
        openaiKeySet:
          type: boolean
        storageConnectionSet:
          type: boolean
      required:
        - docIntelEndpoint
        - openaiEndpoint
        - openaiDeployment
        - docIntelKeySet
        - openaiKeySet
        - storageConnectionSet
    AzureSettingsInput:
      type: object
      properties:
        docIntelEndpoint:
          type: string
        docIntelKey:
          type: string
        openaiEndpoint:
          type: string
        openaiKey:
          type: string
        openaiDeployment:
          type: string
        storageConnectionString:
          type: string
    UnlockSettingsInput:
      type: object
      properties:
        password:
          type: string
          minLength: 1
      required:
        - password
    UnlockSettingsResult:
      type: object
      properties:
        token:
          type: string
        expiresAt:
          type: string
      required:
        - token
        - expiresAt
    DocumentUpload:
      type: object
      properties:
        file:
          type: string
          format: binary
      required:
        - file
    ImportedLoanFields:
      type: object
      properties:
        name:
          type: string
          nullable: true
        lender:
          type: string
          nullable: true
        principal:
          type: number
          nullable: true
        downPayment:
          type: number
          nullable: true
        interestRate:
          type: number
          nullable: true
        primeSpread:
          type: number
          nullable: true
        statedPrimeRate:
          type: number
          nullable: true
        amortizationYears:
          type: number
          nullable: true
        termYears:
          type: number
          nullable: true
        startDate:
          type: string
          nullable: true
        paymentFrequency:
          type: string
          enum: [monthly, semi-monthly, bi-weekly, weekly]
          nullable: true
        paymentAmount:
          type: number
          nullable: true
        interestOnlyMonths:
          type: number
          nullable: true
        balloonPayment:
          type: number
          nullable: true
        securityDescription:
          type: string
          nullable: true
    ImportedLeaseFields:
      type: object
      properties:
        name:
          type: string
          nullable: true
        lessor:
          type: string
          nullable: true
        assetDescription:
          type: string
          nullable: true
        assetType:
          type: string
          enum: [vehicle, equipment, office_commercial, other]
          nullable: true
        startDate:
          type: string
          nullable: true
        termMonths:
          type: number
          nullable: true
        monthlyPayment:
          type: number
          nullable: true
        downPayment:
          type: number
          nullable: true
        interestRate:
          type: number
          nullable: true
        fairValue:
          type: number
          nullable: true
        economicLifeYears:
          type: number
          nullable: true
        buyoutAmount:
          type: number
          nullable: true
        transferOfOwnership:
          type: boolean
          nullable: true
        bargainPurchaseOption:
          type: boolean
          nullable: true
        specializedAsset:
          type: boolean
          nullable: true
        paymentAtBeginning:
          type: boolean
          nullable: true
        rentableSquareFeet:
          type: number
          nullable: true
        rentSteps:
          type: array
          items:
            $ref: "#/components/schemas/RentStep"
          nullable: true
        tenantImprovementAllowance:
          type: number
          nullable: true
        freeRentMonths:
          type: number
          nullable: true
        camAnnualPerSquareFoot:
          type: number
          nullable: true
        camMonthly:
          type: number
          nullable: true
        percentageRentNote:
          type: string
          nullable: true
        fieldNotes:
          type: object
          additionalProperties:
            type: string
          nullable: true
        estimates:
          allOf:
            - $ref: "#/components/schemas/LeaseEstimates"
          nullable: true
    LeaseEstimates:
      type: object
      properties:
        economicLifeYears:
          type: number
          nullable: true
        fairValue:
          type: number
          nullable: true
        interestRate:
          type: number
          nullable: true
        reasoning:
          type: string
          nullable: true
    RentStep:
      type: object
      properties:
        fromYear:
          type: integer
        toYear:
          type: integer
        monthlyRent:
          type: number
        annualRatePerSquareFoot:
          type: number
          nullable: true
      required:
        - fromYear
        - toYear
        - monthlyRent
    DocumentImportResult:
      type: object
      properties:
        classification:
          type: string
          enum: [loan, lease, other]
        confidence:
          type: number
        reasoning:
          type: string
        documentBlob:
          type: string
          nullable: true
        documentName:
          type: string
          nullable: true
        loan:
          allOf:
            - $ref: "#/components/schemas/ImportedLoanFields"
          nullable: true
        lease:
          allOf:
            - $ref: "#/components/schemas/ImportedLeaseFields"
          nullable: true
      required:
        - classification
        - confidence
        - reasoning
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error