# lib/api-zod/src/generated/types/azureSettingsStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AzureSettingsStatus {
  docIntelEndpoint: string;
  openaiEndpoint: string;
  openaiDeployment: string;
  docIntelKeySet: boolean;
  openaiKeySet: boolean;
  storageConnectionSet: boolean;
}